+++ autosubmit/job/job.py
@@ -243,7 +243,7 @@ class Job(object):
         self._local_logs = ('', '')
         self._remote_logs = ('', '')
         self.script_name = self.name + ".cmd"
-        self.stat_file = self.script_name[:-4] + "_STAT_"
+        self.stat_file = f"{self.script_name[:-4]}_STAT_"
         self._status = None
         self.status = status
         self.prev_status = status
@@ -387,7 +387,7 @@ class Job(object):
         self.start_time_timestamp = time.time()
         self.processors_per_node = ""
         self.script_name = self.name + ".cmd"
-        self.stat_file = self.script_name[:-4] + "_STAT_"
+        self.stat_file = f"{self.script_name[:-4]}_STAT_"
         self.reservation = ""
         self.current_checkpoint_step = 0
         self.max_checkpoint_step = 0
@@ -2921,18 +2921,18 @@ class WrapperJob(Job):
             remote_log_dir = self._platform.get_remote_log_dir()
             # PREPARE SCRIPT TO SEND
             # When an inner_job is running? When the job has an _STAT file
-            command = textwrap.dedent("""
-            cd {1}
-            for job in {0}
+            command = textwrap.dedent(f"""
+            cd {str(remote_log_dir)}
+            for job in {str(not_finished_jobs_names)}
             do
-                if [ -f "${{job}}_STAT" ]
+                if [ -f "${{job}}_STAT_{self.fail_count}" ]
                 then
-                        echo ${{job}} $(head ${{job}}_STAT)
+                        echo ${{job}} $(head ${{job}}_STAT_{self.fail_count})
                 else
                         echo ${{job}}
                 fi
             done
-            """).format(str(not_finished_jobs_names), str(remote_log_dir), '\n'.ljust(13))
+            """)
 
             log_dir = Path(str(self._tmp_path) + f'/LOG_{self.expid}')
             multiple_checker_inner_jobs = Path(log_dir / "inner_jobs_checker.sh")