+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### 4.1.14: Dependencies bug fixes, regression tests
+
+- Fixed issues with wrapper not updating status correctly. [#2274](https://github.com/BSC-ES/autosubmit/issues/2274) [PR](https://github.com/BSC-ES/autosubmit/pull/2327)
+
 ### 4.1.13: Dependencies bug fixes, regression tests
 
 - Fixed issues with dependencies not being correctly set or not being pruned. [#2184](https://github.com/BSC-ES/autosubmit/issues/2184) [PR](https://github.com/BSC-ES/autosubmit/pull/2241)